Returner REVUDERINGmappe i sanity dersom behandlingen er en revurdering
